requirements.txt
fastapi==0.103.2
pydantic==2.4.2
uvicorn==0.23.2
websockets==11.0.3
python-dotenv==1.0.0
SQLAlchemy==2.0.22
asyncpg==0.28.0
psycopg2==2.9.9
alembic==1.12.1
bcrypt==4.0.1
python-jose==3.3.0